moved folders & fixed index.js file
