+++ server/src/models/index.js
@@ -1,7 +1,46 @@
-const sequelize = require('../config/database'); // Import the Sequelize instance
-const User = require('./User'); // Import User model
-const Event = require('./Event'); // Import Event model
-const Rsvp = require('./Rsvp'); // Import Rsvp model
+'use strict';
+
+const fs = require('fs');
+const path = require('path');
+const Sequelize = require('sequelize');
+const process = require('process');
+const basename = path.basename(__filename);
+const env = process.env.NODE_ENV || 'development';
+const config = require('../config/config.js')[env]; // Using config.js instead of config.json
+const db = {};
+
+// Initialize Sequelize connection
+let sequelize;
+if (config.use_env_variable) {
+  sequelize = new Sequelize(process.env[config.use_env_variable], config);
+} else {
+  sequelize = new Sequelize(config.database, config.username, config.password, config);
+}
+
+// Dynamically import models from the current directory
+fs
+  .readdirSync(__dirname)
+  .filter(file => {
+    return (
+      file.indexOf('.') !== 0 &&
+      file !== basename &&
+      file.slice(-3) === '.js' &&
+      file.indexOf('.test.js') === -1
+    );
+  })
+  .forEach(file => {
+    const model = require(path.join(__dirname, file))(sequelize, Sequelize.DataTypes);
+    db[model.name] = model;
+  });
+
+// Import models explicitly
+const User = require('./User');
+const Event = require('./Event');
+const Rsvp = require('./Rsvp');
+
+db.User = User;
+db.Event = Event;
+db.Rsvp = Rsvp;
 
 // Define Associations
 User.hasMany(Event, { foreignKey: 'created_by', onDelete: 'CASCADE' });
@@ -13,14 +52,19 @@ Event.hasMany(Rsvp, { foreignKey: 'event_id', onDelete: 'CASCADE' });
 Rsvp.belongsTo(User, { foreignKey: 'user_id', onDelete: 'CASCADE' });
 Rsvp.belongsTo(Event, { foreignKey: 'event_id', onDelete: 'CASCADE' });
 
-// Sync models with the database (optional but useful)
+// Sync models with the database
 const syncDatabase = async () => {
   try {
     await sequelize.sync({ alter: true }); // Ensure tables are updated without losing data
-    console.log('Database synchronized successfully.');
+    console.log('✅ Database synchronized successfully.');
   } catch (error) {
-    console.error('Error syncing database:', error);
+    console.error('❌ Error syncing database:', error);
   }
 };
 
-module.exports = { sequelize, User, Event, Rsvp, syncDatabase };
+// Attach Sequelize and the sync function to the db object
+db.sequelize = sequelize;
+db.Sequelize = Sequelize;
+db.syncDatabase = syncDatabase;
+
+module.exports = db;
