+++ tsconfig.json
@@ -22,10 +22,10 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "exactOptionalPropertyTypes": true, /* Interpret optional property types as written, rather than adding 'undefined'. */,
-    "noImplicitReturns": true, /* Enable error reporting for codepaths that do not explicitly return in a function. */,
-    "noImplicitOverride": true, /* Ensure overriding members in derived classes are marked with an override modifier. */,
-    "useUnknownInCatchVariables": true, /* Ensure 'unknown' is used for catch clause variables. */
+    "exactOptionalPropertyTypes": true /* Interpret optional property types as written, rather than adding 'undefined'. */,
+    "noImplicitReturns": true /* Enable error reporting for codepaths that do not explicitly return in a function. */,
+    "noImplicitOverride": true /* Ensure overriding members in derived classes are marked with an override modifier. */,
+    "useUnknownInCatchVariables": true /* Ensure 'unknown' is used for catch clause variables. */,
     "rootDir": "./",
     "outDir": "dist"
   },
